wip(1584): Subscribers refactored, created generic hook for data tables and changed approach how data is handled in tables
