+++ dashboard/local.env
@@ -1,0 +1,2 @@
+REACT_APP_ENABLE_SOCIAL_AUTH=true
+REACT_APP_RECAPTCHA_SITE_KEY=
